Move contribute thanks page to /modules/contribute
